replace "unsigned long long" for SgfcNodeTraits and SgfcPropertyTraits with uint64_t (#51)

+++ include/SgfcNodeTraits.h
@@ -19,6 +19,9 @@
 // Project includes (generated)
 #include "SgfcPlusPlusExport.h"
 
+// C++ Standard Library includes
+#include <cstdint>
+
 /// @brief file
 
 namespace LibSgfcPlusPlus
@@ -51,7 +54,7 @@ namespace LibSgfcPlusPlus
   /// @note Using an unsigned numerical type as the underlying type of this
   /// typedef is important to make the bitwise operators, specifically the
   /// shift left and shift right operators, unambiguous and safe to use.
-  typedef unsigned long long SgfcNodeTraits;
+  typedef uint64_t SgfcNodeTraits;
 
   /// @brief SgfcNodeTrait enumerates the individual traits that an
   /// ISgfcNode object can have. An ISgfcNode object can have several
